creating reviews.js
